Refactor: useExtendedSearchBar 작성
